requirements.txt
dj-database-url==0.5.0
Django==2.1.3
django-dotenv==1.4.2
djangorestframework==3.9.0
psycopg2==2.7.5
pytz==2018.7
